+++ changes.md
@@ -35,6 +35,9 @@
 - added property `deliveryKey` to type `OrderSetDeliveryCustomFieldAction`
 - added property `deliveryKey` to type `OrderSetDeliveryCustomTypeAction`
 - added property `deliveryKey` to type `OrderSetDeliveryItemsAction`
+- added property `sku` to type `ProductRemovePriceAction`
+- added property `variantId` to type `ProductRemovePriceAction`
+- added property `price` to type `ProductRemovePriceAction`
 - added property `myBusinessUnitAssociateRoleOnCreation` to type `BusinessUnitConfiguration`
 </details>